// packages/cookie/src/types.ts
export interface CookieParseOptions {
  /**
   * Specifies a function that will be used to decode a cookie's value. Since
   * the value of a cookie has a limited character set (and must be a simple
   * string), this function can be used to decode a previously-encoded cookie
   * value into a JavaScript string or other object.
   *
   * The default function is the global `decodeURIComponent`, which will decode
   * any URL-encoded sequences into their byte representations.
   *
   * Note* if an error is thrown from this function, the original, non-decoded
   * cookie value will be returned as the cookie's value.
   */
  decode?: (value: string) => string;
  /**
   * Custom function to filter parsing specific keys.
   */
  filter?: (key: string) => boolean;
}

export interface SetCookieParseOptions {
  /**
   * Custom decode function to use on cookie values.
   *
   * By default, `decodeURIComponent` is used.
   *
   * **Note:** If decoding fails, the original (undecoded) value will be used
   */
  decode?: false | ((value: string) => string);
}

export interface SetCookie {
  /**
   * Cookie name
   */
  name: string;

  /**
   * Cookie value
   */
  value: string;

  /**
   * Cookie path
   */
  path?: string | undefined;

  /**
   * Absolute expiration date for the cookie
   */
  expires?: Date | undefined;

  /**
   * Relative max age of the cookie in seconds from when the client receives it (integer or undefined)
   *
   * Note: when using with express's res.cookie() method, multiply maxAge by 1000 to convert to milliseconds
   */
  maxAge?: number | undefined;

  /**
   * Domain for the cookie,
   * May begin with "." to indicate the named domain or any subdomain of it
   */
  domain?: string | undefined;

  /**
   * Indicates that this cookie should only be sent over HTTPs
   */
  secure?: boolean | undefined;

  /**
   * Indicates that this cookie should not be accessible to client-side JavaScript
   */
  httpOnly?: boolean | undefined;

  /**
   * Indicates a cookie ought not to be sent along with cross-site requests
   */
  sameSite?: true | false | "lax" | "strict" | "none" | undefined;

  [key: string]: unknown;
}